[IT-255] Synapse External bucket PR example
This is an example PR (do not merge!).  It is a template to show
how to get the CI system to automatically provision a Synapse External
bucket.

Bucket creation workflow:
  1. Scientist create a PR similar to this one.  Create a new file
     with a unique stack_name and updated parameters.
  2. Have someone review and approve the PR.
  3. Merge the PR
  4. The CI system will auto provision the bucket
  5. An email will be sent to the bucket owner email address with
     the new bucket info.

The newly created bucket will contain the following tags:

 key: aws:cloudformation:stack-id
 value: arn:aws:cloudformation:us-east-1:<AWS account ID>:stack/GatesKI-TestStudy/375ced80-78b9-11e8-bc55-500c286e1a36

 key: aws:cloudformation:stack-name
 value: GatesKI-TestStudy

 key: aws:cloudformation:logical-id
 value: SynapseExternalBucket

 key: OwnerEmail
 value: joe.smith@sagebase.org

Since cloudformation does not allow deleting a buckets with data in it
the bucket owner will need to contact an admin to delete the bucket
on their behalf.