Remove genie ec2
Genie wants to use an AMI built from ubuntu which does not send the
success signal required by the managed EC2 CF template to deploy
instances.  The image would get spun up successfully however CF
fails because it's expecting a success signal which it never
recieves.

Remove provisioning of this genie instance until we find a work
around for this problem.