Setup a sagebionetworks.org domain
Setup a *.sagebionetworks.org ACM certificate for sagebionetworks.org
domain.

This was manually run because it required a manual step of adding DNS
validation to godaddy.  Added to CI so we can make updates thru git
pull requests.

This change is dependent on PR https://github.com/Sage-Bionetworks/aws-infra/pull/97