Add a batch developer group
Due to the limitations of batch policies[1] we cannot scope user access
to specific actions on specific batch jobs.  Therefore we need to provide
broder access to make it usable to a batch developer.  Adding a batch
developer group for Tom so he can develop and trigger batch jobs from
the AWS console.

Note-this group would have access to all batch resources not just the
ones they created.

[1] https://docs.aws.amazon.com/batch/latest/userguide/batch-supported-iam-actions-resources.html