Change Sceptre Hooks to before_launch (#217)
* Fix S3 URL Formatting

* Combine Hooks into before_launch

IMPORTANT NOTE: some of these config files only had a before_update
hook. In this pull request, I'm changing those to before_launch hooks as
well. I'm not confident that this won't cause any undesired effects, so
I whoever is reviewing this should make a special effort to ensure that
those changes won't be a problem.